+++ productAppBE/src/main/java/com/hellokoding/springboot/restful/controller/OrgAPI.java
@@ -30,7 +30,7 @@ public class OrgAPI {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Org> findById(@PathVariable Long id) {
+    public ResponseEntity<Org> findById(@PathVariable Integer id) {
         Optional<Org> stock = orgService.findById(id);
         if (!stock.isPresent()) {
             log.error("Id " + id + " is not existed");
@@ -40,7 +40,7 @@ public class OrgAPI {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Org> update(@PathVariable Long id, @Valid @RequestBody Org org) {
+    public ResponseEntity<Org> update(@PathVariable Integer id, @Valid @RequestBody Org org) {
         if (!orgService.findById(id).isPresent()) {
             log.error("Id " + id + " is not existed");
             ResponseEntity.badRequest().build();
@@ -49,7 +49,7 @@ public class OrgAPI {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity delete(@PathVariable Long id) {
+    public ResponseEntity delete(@PathVariable Integer id) {
         if (!orgService.findById(id).isPresent()) {
             log.error("Id " + id + " is not existed");
             ResponseEntity.badRequest().build();
